# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BSY MIS Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sanstha/{id}:
    get:
      operationId: getSanstha
      tags: [sanstha]
      summary: Get sanstha by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sanstha
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sanstha"
  /sanstha/{id}/students:
    get:
      operationId: getSansthaStudents
      tags: [sanstha]
      summary: Get students enrolled in sanstha
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Beneficiary"
...
components:
  schemas:
...
    Beneficiary:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        childName:
          type: string
        district:
          type: string
        approvalDate:
          type: string
        sansthaId:
          type: ["integer", "null"]
        sansthaName:
          type: ["string", "null"]
        beneficiaryCategory:
          type: string
        monthlyAmount:
          type: number
        status:
          type: string
          enum: [active, inactive, aged_out, suspended]
        aadhaarSeeded:
          type: boolean
        kycDone:
          type: boolean
        eligibilityRisk:
          type: boolean
        renewalDue:
          type: boolean
      required: [id, applicationId, childName, district, approvalDate, beneficiaryCategory, monthlyAmount, status, aadhaarSeeded, kycDone, eligibilityRisk, renewalDue]
...
    Sanstha:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        district:
          type: string
        maxStudents:
          type: integer
        enrolledCount:
          type: integer
        contactPerson:
          type: string
        mobile:
          type: string
        status:
          type: string
          enum: [active, inactive]
      required: [id, name, district, maxStudents, enrolledCount, contactPerson, mobile, status]
...